extract run inventory routes
